pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = [
    "Shoxrux Yuldashov <deagleshox@mail.ru>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "5.0"
django-jet-reboot = "1.3.9"
python-decouple = "3.8"
psycopg2-binary = "2.9.9"
pillow = "10.4.0"
easy-thumbnails = "2.10"
markdown = "3.7"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
django-filter = "24.3"
requests = "2.32.3"
pytest-django = "4.9.0"
gunicorn = "21.2.0"
redis = "5.2.0"
celery = "5.4.0"
django-cors-headers = "4.6.0"
coreapi = "2.3.3"
pyyaml = "6.0.2"
djoser = "2.3.1"
drf-spectacular = "^0.27.2"


[tool.poetry.group.dev.dependencies]
black = "24.8.0"
isort = "5.13.2"
flake8 = "7.1.1"
flake8-pyproject = "1.2.3"
djlint = "1.34.2"
mypy = "1.11.2"
pylint = "3.2.6"

[tool.poetry.group.test.dependencies]
pytest = "8.3.2"
faker = "28.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
line-length = 79
max-complexity = 5
per-file-ignores = "__init__.py:F401"
extend-ignore = ["E203", "F403"]
exclude = [
    ".git/",
    "__pycache__/",
    "**/migrations/",
    "manage.py",
    "**settings/*.py",
    "env/",
    "venv/",
    ".env/",
    ".venv/",
    "tests/",
    ".pytest_cache",
    ".vscode/",
    ".idea/"
]

[tool.flake8.ignore]
W503 = true
F811 = true
R503 = true
R504 = true
R505 = true

[tool.black]
line-length = 79
include = '\.pyi?$'
extend-exclude = '''
/(
    migrations
  | __pycache__
  | .pytest_cache
  | .vscode
  | .idea
)/
'''

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_local_folder = [
    "core",
    "api",
    "applications",
    "users",
]
